+++ src/main/java/org/management/asset/services/RiskScenarioService.java
@@ -14,7 +14,7 @@ public interface RiskScenarioService {
 
     List<RiskScenario> getRiskScenarios();
 
-    Long getRiskScenariosCounter();
+    Integer getRiskScenariosCounter();
 
     RiskScenarioResponseDTO saveRiskScenario(RiskScenarioRequestDTO riskScenarioRequest);
